Removing auto setting for dtrace config option on jdk15+
In JDK15 and up, "--enable-dtrace=auto" does not appear to be a valid
config argument for jdk builds. In the adoptium pipeline builds, using
"--enable-dtrace" seems to work fine, so here's a fix along those
lines.

Second, we shouldn't use the dtrace option on Windows, and Linux needs a
dependancy to build with dtrace enabled, so I've included fixes for
those issues too.

Lastly, the bundling of freetype appears to add a dependency on sdt.h,
which cannot be found. Since normal jenkins pipeline builds do not
bundle freetype on linux, it seems sensible to skip bundling in this
case as well.

Signed-off-by: Adam Farley <adfarley@redhat.com>