Remove dtrace from all linux builds as it doesn't work
On JDK15 and up, attempting (and failing) to enable dtrace will
cause the build to fail during the configure step, so I'm excluding
the dtrace enabling code for now.

On JDK14 and below, attempting (and failing) to enable dtrace will
not cause the build to fail, but the error message isn't obvious
unless you read the entire configure output, so I'm excluding it
there for consistency.

Full details on "why" dtrace cannot be enabled in a github actions
build has been added into the code as a todo comment. This can be
resolved if and when it reaches the top of someone's priority stack.

Addendum: openj9 jdk11 builds stopped working with jdk10 boot jdks,
so I'm including a small change that builds jdk11 on jdk11 boot jdks.

Signed-off-by: Adam Farley <adfarley@redhat.com>